web/mooctracker/requirements.txt
Django==1.6.5
argparse==1.2.1
django-social-auth==0.7.28
djangorestframework==2.3.13
httplib2==0.9
oauth2==1.5.211
python-openid==2.2.5
wsgiref==0.1.2
